telegram_chatbot/requirements.txt
docx==0.2.4
langchain==0.0.340
openai==0.28.0
tiktoken==0.5.1
PyPDF2==3.0.1
python_docx==1.1.0
pysqlite3-binary
chromadb==0.4.0
telebot==0.0.5
python-dotenv
moviepy
pydub